+++ android/src/main/kotlin/com/f2fk/geofence_foreground_service/interfaces/Builder.kt
@@ -1,0 +1,5 @@
+package com.f2fk.geofence_foreground_service.interfaces
+
+interface Builder<T> {
+    fun build() : T
+}