+++ lib/geofence_foreground_service_method_channel.dart
@@ -7,22 +7,25 @@ import 'package:flutter/widgets.dart';
 import 'package:geofence_foreground_service/constants/geofence_event_type.dart';
 import 'package:geofence_foreground_service/models/zone.dart';
 
+import 'constants/api_methods.dart';
 import 'constants/json_keys.dart';
 import 'geofence_foreground_service_platform_interface.dart';
 import 'models/background_task_handlers.dart';
 import 'models/notification_icon_data.dart';
 
 /// An implementation of [GeofenceForegroundServicePlatform] that uses method channels.
-class MethodChannelGeofenceForegroundService
-    extends GeofenceForegroundServicePlatform {
+class MethodChannelGeofenceForegroundService extends GeofenceForegroundServicePlatform {
   /// The method channel used to interact with the native platform.
   @visibleForTesting
-  final foregroundChannel = const MethodChannel(
-      'ps.byshy.geofence/foreground_geofence_foreground_service');
+  final foregroundChannel =
+      const MethodChannel('ps.byshy.geofence/foreground_geofence_foreground_service');
 
   @visibleForTesting
-  final backgroundChannel = const MethodChannel(
-      "ps.byshy.geofence/background_geofence_foreground_service");
+  final backgroundChannel =
+      const MethodChannel("ps.byshy.geofence/background_geofence_foreground_service");
+
+  Future<T?> _invokeMethod<T>(String method, [dynamic arguments]) =>
+      foregroundChannel.invokeMethod(method, arguments);
 
   /// This method is used to start the geofencing foreground service
   @override
@@ -61,11 +64,7 @@ class MethodChannelGeofenceForegroundService
       bool didStart = false;
 
       try {
-        didStart = await foregroundChannel.invokeMethod<bool?>(
-              'startGeofencingService',
-              data,
-            ) ??
-            false;
+        return await _invokeMethod<bool?>(ApiMethods.startGeofencingService, data) ?? false;
       } catch (e) {
         log(
           e.toString(),
@@ -91,9 +90,7 @@ class MethodChannelGeofenceForegroundService
 
       return backgroundTriggerHandler(
         call.arguments['ps.byshy.geofence.ZONE_ID'],
-        inputData == null
-            ? GeofenceEventType.unKnown
-            : (jsonDecode(inputData) as int).toGeofenceEventType(),
+        GeofenceEventType.findById(jsonDecode(inputData) as int?),
       );
     });
 
@@ -106,9 +103,7 @@ class MethodChannelGeofenceForegroundService
     bool didStop = false;
 
     try {
-      didStop = await foregroundChannel
-              .invokeMethod<bool?>('stopGeofencingService') ??
-          false;
+      return await _invokeMethod<bool?>(ApiMethods.stopGeofencingService) ?? false;
     } catch (e) {
       log(
         e.toString(),
@@ -125,9 +120,7 @@ class MethodChannelGeofenceForegroundService
     bool isServiceRunning = false;
 
     try {
-      isServiceRunning = await foregroundChannel
-              .invokeMethod<bool?>('isForegroundServiceRunning') ??
-          false;
+      return await _invokeMethod<bool?>(ApiMethods.isForegroundServiceRunning) ?? false;
     } catch (e) {
       log(
         e.toString(),
@@ -146,11 +139,7 @@ class MethodChannelGeofenceForegroundService
     bool isGeofenceAdded = false;
 
     try {
-      isGeofenceAdded = await foregroundChannel.invokeMethod<bool?>(
-            'addGeofence',
-            zone.toJson(),
-          ) ??
-          false;
+      return await _invokeMethod<bool?>(ApiMethods.addGeofence, zone.toJson()) ?? false;
     } catch (e) {
       log(
         e.toString(),
@@ -169,12 +158,7 @@ class MethodChannelGeofenceForegroundService
     bool isGeofenceRemoved = false;
 
     try {
-      isGeofenceRemoved = await foregroundChannel.invokeMethod<bool?>(
-            'removeGeofence',
-            {
-              JsonKeys.zoneId: zoneId,
-            },
-          ) ??
+      return await _invokeMethod<bool?>(ApiMethods.removeGeofence, {JsonKeys.zoneId: zoneId}) ??
           false;
     } catch (e) {
       log(
@@ -192,9 +176,7 @@ class MethodChannelGeofenceForegroundService
     bool areAllAreasRemoved = false;
 
     try {
-      areAllAreasRemoved =
-          await foregroundChannel.invokeMethod<bool?>('removeAllGeoFences') ??
-              false;
+      return await _invokeMethod<bool?>(ApiMethods.removeAllGeoFences) ?? false;
     } catch (e) {
       log(
         e.toString(),